# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ventrix API",
    "version": "1.0.0",
    "description": "API de Ventrix POS y Ventrix Contable. Generada automáticamente desde src/routes (`npm run docs:openapi`), así que refleja siempre las rutas, roles y validaciones que hay en el código.\n\nLa sesión usa un access token corto (cabecera `Authorization: Bearer`) y una cookie httpOnly de refresh. Todo es multi-tenant: el negocio sale del token, nunca del cuerpo de la petición."
  },
  "servers": [
    {
      "url": "https://api.ventrix.lat/api/v1",
      "description": "Producción"
    },
    {
      "url": "http://localhost:4000/api/v1",
      "description": "Local"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token de un usuario (POS o Contable)."
      },
      "sellerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Sesión del portal de vendedoras (/seller), separada de la de usuarios."
      }
    },
    "schemas": {
      "RespuestaOk": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "description": "Contenido de la respuesta"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Mensaje del error"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/credits": {
      "get": {
        "tags": [
          "credit"
        ],
        "summary": "GET /api/v1/credits",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "credit"
        ],
        "summary": "POST /api/v1/credits",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "totalAmount": {
                    "type": "number",
                    "description": "El monto del crédito debe ser mayor a 0",
                    "minimum": 0.01
                  }
                },
                "required": [
                  "totalAmount"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/credits/{id}": {
      "get": {
        "tags": [
          "credit"
        ],
        "summary": "GET /api/v1/credits/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v1/credits/{id}/payments": {
      "post": {
        "tags": [
          "credit"
        ],
        "summary": "POST /api/v1/credits/{id}/payments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Monto inválido",
                    "minimum": 0.01
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "CASH",
                      "TRANSFER",
                      "NEQUI",
                      "DAVIPLATA",
                      "CARD"
                    ]
                  },
                  "paymentAccountId": {
                    "type": "string"
                  },
                  "installmentId": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Cuota inválida"
                  }
                },
                "required": [
                  "amount"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/v1/notifications/test": {
      "post": {
        "tags": [
          "notification"
        ],
        "summary": "Envía un push de PRUEBA a los dispositivos suscritos del usuario. Sirve para que confirme, en un toque, que las notificaciones le llegan al celular.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/seller/push/test": {
      "post": {
        "tags": [
          "seller"
        ],
        "summary": "POST /seller/push/test → prueba en un toque, para que confirme que le llegan.",
        "security": [
          {
            "sellerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/supplier-credits": {
      "get": {
        "tags": [
          "supplierCredit"
        ],
        "summary": "GET /api/v1/supplier-credits",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/supplier-credits/{id}": {
      "get": {
        "tags": [
          "supplierCredit"
        ],
        "summary": "GET /api/v1/supplier-credits/{id}",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v1/supplier-credits/{id}/payments": {
      "post": {
        "tags": [
          "supplierCredit"
        ],
        "summary": "POST /api/v1/supplier-credits/{id}/payments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "El monto debe ser mayor a 0"
                  },
                  "paymentMethod": {
                    "type": "string",
                    "enum": [
                      "CASH",
                      "TRANSFER",
                      "NEQUI",
                      "DAVIPLATA",
                      "CARD"
                    ]
                  },
                  "paymentAccountId": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                },
                "required": [
                  "amount"
                ]
              }
            }
          }
        }
      }
    },
    "/api/v1/supplier-credits/import": {
      "post": {
        "tags": [
          "supplierCredit"
        ],
        "summary": "Abonar a un proveedor: paga parte del saldo → SALE plata de la caja. ── Importación masiva (admite ?dryRun=true para la vista previa) ───────────── Sirve para cargar de una las facturas que el negocio ya tiene pendientes con sus proveedores, en vez de teclearlas una por una al empezar a usar Ventrix",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/supplier-credits/import-template": {
      "get": {
        "tags": [
          "supplierCredit"
        ],
        "summary": "── Plantilla (sin sesión: es un archivo en blanco, no lleva datos de nadie) ──",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suppliers": {
      "get": {
        "tags": [
          "supplier"
        ],
        "summary": "── Standard CRUD ────────────────────────────────────────────────────────────",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "supplier"
        ],
        "summary": "CASHIER puede crear un proveedor nuevo al vuelo (ej. registrando una compra de un proveedor que aún no existe en el sistema), pero no editar/eliminar uno existente — eso sigue siendo de ADMIN/SUPERVISOR.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaOk"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token vencido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}